Home

PlugAndPlot.jl

GUI to create statistical plots from a DataFrame, using DataFrames, QML and StatPlots

Installation

The package is not yet registered in Julia's package registry, and so it must be installed using

Pkg.clone("https://github.com/piever/PlugAndPlot.jl.git")

Documentation

PlugAndPlot.build_window — Function.

build_window(; kwargs...)

Starts the GUI asking for a suitable csv file.

source

build_window(datafile::AbstractString; kwargs...)

Reads a csv file and starts build_window on the corresponding DataFrame

source

build_window(dataset::AbstractDataFrame; nbox = 5)

Creates a GUI to analyze a DataFrame interactively. Data can be selected either on continuous columns, with SpinBoxes or on discrete columns with checkboxes, provided there are less than nbox entries.

source