Getting Started

Getting started

Loading the data

To load the data, simply type the following command in the Julia REPL:

using PlugAndPlot
build_window()

you will then be asked to choose for a csv file with your data. It is important to store your data as a tabular dataset. In particular

Data is represented as a series of columns with equal length, with values separated by commas

The first row is assumed to be the header

Missing data is not supported yet

PlugAndPlot.build_window — Function.

build_window(; kwargs...)

Starts the GUI asking for a suitable csv file.

source

build_window(datafile::AbstractString; kwargs...)

Reads a csv file and starts build_window on the corresponding DataFrame

source

build_window(dataset::AbstractDataFrame; nbox = 5)

Creates a GUI to analyze a DataFrame interactively. Data can be selected either on continuous columns, with SpinBoxes or on discrete columns with checkboxes, provided there are less than nbox entries.

source